Explicitly specify tox envs instead of using tox-travis
tox-travis wasn't running the lint env, so get rid
of it in favor of explictly specifying the TOXENV variable
in .travis.yml.

This makes the Travis config more explicit whilst simplifying
the tox config.

One downside: Build stages doesn't seem to work with this
approach, so we use Travis' "normal" PyPI deployment.
We can look into this later