Add Metric.record and remove date param from Metric.save
In retrospect, adding the `date` param in #43 was a bad idea.
As @pattisdr pointed out in
https://github.com/sloria/django-elasticsearch-metrics/pull/5#discussion_r205828049,
a user could pass an index and a date that are in conflict.

With this change, Metric.save will choose the index based on the
value of the timestamp field.

This also adds a convenience method, Metric.record, that allows
a timestamp to be passed

```python
MyMetric.record(timestamp=..., my_field=...)
```

This improves the ergonomics because a user typically does not
need instances of a metric.